Crud methods with endpoints

+++ src/main/java/dat3/car/cars/dto/ReservationRequest.java
@@ -1,9 +1,6 @@
 package dat3.car.cars.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import dat3.car.cars.entity.Car;
-import dat3.car.cars.entity.Member;
-import dat3.car.cars.entity.Reservation;
 import lombok.*;
 
 import java.time.LocalDate;
@@ -16,8 +13,8 @@ import java.time.LocalDate;
 public class ReservationRequest {
 
     // Attributes
-    int carId;
+    Long carId;
     String userName;
-    @JsonFormat(pattern = "yyyy-MM-dd",shape = JsonFormat.Shape.STRING)
+    @JsonFormat(pattern = "yyyy-MM-dd", shape = JsonFormat.Shape.STRING)
     LocalDate date;
 }
